Pull out common event dispatcher.
